invoice-csv-tool/requirements.txt
typing
pyyaml

google-api-core
google-cloud-documentai
google-cloud-storage

Flask
gunicorn
pandas
